Move DatabaseReader and DatabaseWriter to KrepostLib.Storage namespace
